+++ package.json
@@ -1,4 +1,7 @@
 {
+  "dependencies": {
+    "esbuild": "^0.23.0"
+  },
   "devDependencies": {
     "husky": "^9.0.11",
     "lint-staged": "^15.2.7",
@@ -13,8 +16,8 @@
   },
   "scripts": {
     "prepare": "husky",
-    "test": "npx http-server -o \"site?demo=1\" -c-1",
-    "publish": "npx esbuild \"site/*\" \"site/schemas/*\" \"site/scripts/script.js\" --minify --bundle --loader:.jpg=copy --loader:.html=copy --loader:.ico=copy --loader:.png=copy --loader:.webmanifest=copy --loader:.proto=copy --loader:.json=empty --outdir=\"min\" --analyze"
+    "publish": "node .esbuild.mjs",
+    "test": "npx http-server -o \"site?demo=1\" -c-1"
   },
   "lint-staged": {
     "**/*.{js,md,html,css}": "prettier --write"
