+++ site/styles/load-backup.css
@@ -1,0 +1,7 @@
+/* ---Load Modal--- */
+#load-modal h2 {
+  padding-top: 0.5em;
+}
+#file-input {
+  margin: 10px 0;
+}
